refactor(plans): move usePlanNameTranslation to a better location
refs #1425
