+++ src/components/Pricing/OrderSummary.tsx
@@ -1,47 +1,9 @@
 import { Box, Divider, Flex, Text, VStack } from '@chakra-ui/react'
-import { useCallback, useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
-import { usePlanTranslations, usePlans } from './Plans'
+import { usePlanNameTranslator } from './Plans'
 
 type OrderSummaryProps = {
   checkout: any
-}
-
-type NormalizeFn = (value?: string | null) => string | undefined
-
-type TranslateFn = (rawName?: string | null, fallback?: string | null) => string | undefined
-
-const normalizeName: NormalizeFn = (value) => value?.trim().toLowerCase() || undefined
-
-export const usePlanNameTranslator = () => {
-  const { data: plans } = usePlans()
-  const planTranslations = usePlanTranslations()
-
-  const planTitleByName = useMemo(() => {
-    if (!plans?.length) return {}
-
-    return plans.reduce<Record<string, string>>((acc, plan) => {
-      const normalizedPlanName = normalizeName(plan.name)
-      if (normalizedPlanName) {
-        acc[normalizedPlanName] = planTranslations[plan.id]?.title || plan.name
-      }
-      return acc
-    }, {})
-  }, [plans, planTranslations])
-
-  const translatePlanName = useCallback<TranslateFn>(
-    (rawName, fallback) => {
-      const normalizedName = normalizeName(rawName)
-      if (normalizedName && planTitleByName[normalizedName]) {
-        return planTitleByName[normalizedName]
-      }
-      if (fallback) return fallback
-      return rawName ?? undefined
-    },
-    [planTitleByName]
-  )
-
-  return translatePlanName
 }
 
 export const OrderSummary = ({ checkout }: OrderSummaryProps) => {
